+++ extras/test/cores/coreapi/Arduino.h
@@ -1,0 +1,10 @@
+// StreamUtils - github.com/bblanchon/ArduinoStreamUtils
+// Copyright Benoit Blanchon 2019-2023
+// MIT License
+
+#pragma once
+
+#include "Common.h"
+#include "Print.h"
+#include "Stream.h"
+#include "WString.h"