+++ extras/test/cores/coreapi/Common.h
@@ -1,0 +1,13 @@
+// StreamUtils - github.com/bblanchon/ArduinoStreamUtils
+// Copyright Benoit Blanchon 2019-2023
+// MIT License
+
+#pragma once
+
+#include <time.h>
+
+inline unsigned long millis() {
+  return static_cast<unsigned long>(time(NULL));
+}
+
+inline void yield() {}