+++ extras/test/cores/rp2040/EEPROM.h
@@ -1,0 +1,1 @@
+#include "../esp32/EEPROM.h"